advanced copy

+++ simulation/input_parameters.c
@@ -4,7 +4,7 @@
 // Initialvelocities of all bodies
 double initial_velocities(int n, int d)
 {
-	const int initial_velocities[N][D] = {{0, 0},
+	const double initial_velocities[N][D] = {{0, 0},
 										  {325634, -23630},
 										  {15324, -36234}};
 	return initial_velocities[n][d];
@@ -13,7 +13,7 @@ double initial_velocities(int n, int d)
 // Mass of all bodies
 double initial_masses(int n)
 {
-	const int initial_masses[N] = {542781632,
+	const double initial_masses[N] = {542781632,
 								   12414421,
 								   125252};
 	return initial_masses[n];
